Moved mock statsd client for testing into its own package
