+++ ddtrace/tracer/metrics_test.go
@@ -6,46 +6,14 @@
 package tracer
 
 import (
-	"fmt"
-	"sync"
 	"testing"
 	"time"
 
 	globalinternal "gopkg.in/DataDog/dd-trace-go.v1/internal"
 
 	"github.com/stretchr/testify/assert"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/statsdtest"
 )
-
-type callType int64
-
-const (
-	callTypeGauge callType = iota
-	callTypeIncr
-	callTypeCount
-	callTypeTiming
-)
-
-type testStatsdClient struct {
-	mu          sync.RWMutex
-	gaugeCalls  []testStatsdCall
-	incrCalls   []testStatsdCall
-	countCalls  []testStatsdCall
-	timingCalls []testStatsdCall
-	counts      map[string]int64
-	tags        []string
-	n           int
-	closed      bool
-	flushed     int
-}
-
-type testStatsdCall struct {
-	name     string
-	floatVal float64
-	intVal   int64
-	timeVal  time.Duration
-	tags     []string
-	rate     float64
-}
 
 func withStatsdClient(s globalinternal.StatsdClient) StartOption {
 	return func(c *config) {
@@ -53,193 +21,8 @@ func withStatsdClient(s globalinternal.StatsdClient) StartOption {
 	}
 }
 
-func (tg *testStatsdClient) addCount(name string, value int64) {
-	tg.mu.Lock()
-	defer tg.mu.Unlock()
-	if tg.counts == nil {
-		tg.counts = make(map[string]int64)
-	}
-	tg.counts[name] += value
-}
-
-func (tg *testStatsdClient) Gauge(name string, value float64, tags []string, rate float64) error {
-	return tg.addMetric(callTypeGauge, tags, testStatsdCall{
-		name:     name,
-		floatVal: value,
-		tags:     make([]string, len(tags)),
-		rate:     rate,
-	})
-}
-
-func (tg *testStatsdClient) Incr(name string, tags []string, rate float64) error {
-	tg.addCount(name, 1)
-	return tg.addMetric(callTypeIncr, tags, testStatsdCall{
-		name: name,
-		tags: make([]string, len(tags)),
-		rate: rate,
-	})
-}
-
-func (tg *testStatsdClient) Count(name string, value int64, tags []string, rate float64) error {
-	tg.addCount(name, value)
-	return tg.addMetric(callTypeCount, tags, testStatsdCall{
-		name:   name,
-		intVal: value,
-		tags:   make([]string, len(tags)),
-		rate:   rate,
-	})
-}
-
-func (tg *testStatsdClient) Timing(name string, value time.Duration, tags []string, rate float64) error {
-	return tg.addMetric(callTypeTiming, tags, testStatsdCall{
-		name:    name,
-		timeVal: value,
-		tags:    make([]string, len(tags)),
-		rate:    rate,
-	})
-}
-
-func (tg *testStatsdClient) addMetric(ct callType, tags []string, c testStatsdCall) error {
-	tg.mu.Lock()
-	defer tg.mu.Unlock()
-	copy(c.tags, tags)
-	switch ct {
-	case callTypeGauge:
-		tg.gaugeCalls = append(tg.gaugeCalls, c)
-	case callTypeIncr:
-		tg.incrCalls = append(tg.incrCalls, c)
-	case callTypeCount:
-		tg.countCalls = append(tg.countCalls, c)
-	case callTypeTiming:
-		tg.timingCalls = append(tg.timingCalls, c)
-	}
-	tg.tags = tags
-	tg.n++
-	return nil
-}
-
-func (tg *testStatsdClient) Flush() error {
-	tg.mu.Lock()
-	defer tg.mu.Unlock()
-	tg.flushed++
-	return nil
-}
-
-func (tg *testStatsdClient) Close() error {
-	tg.closed = true
-	return nil
-}
-
-func (tg *testStatsdClient) GaugeCalls() []testStatsdCall {
-	tg.mu.RLock()
-	defer tg.mu.RUnlock()
-	c := make([]testStatsdCall, len(tg.gaugeCalls))
-	copy(c, tg.gaugeCalls)
-	return c
-}
-
-func (tg *testStatsdClient) IncrCalls() []testStatsdCall {
-	tg.mu.RLock()
-	defer tg.mu.RUnlock()
-	c := make([]testStatsdCall, len(tg.incrCalls))
-	copy(c, tg.incrCalls)
-	return c
-}
-
-func (tg *testStatsdClient) CountCalls() []testStatsdCall {
-	tg.mu.RLock()
-	defer tg.mu.RUnlock()
-	c := make([]testStatsdCall, len(tg.countCalls))
-	copy(c, tg.countCalls)
-	return c
-}
-
-func (tg *testStatsdClient) CallNames() []string {
-	tg.mu.RLock()
-	defer tg.mu.RUnlock()
-	var n []string
-	for _, c := range tg.gaugeCalls {
-		n = append(n, c.name)
-	}
-	for _, c := range tg.incrCalls {
-		n = append(n, c.name)
-	}
-	for _, c := range tg.countCalls {
-		n = append(n, c.name)
-	}
-	for _, c := range tg.timingCalls {
-		n = append(n, c.name)
-	}
-	return n
-}
-
-func (tg *testStatsdClient) CallsByName() map[string]int {
-	tg.mu.RLock()
-	defer tg.mu.RUnlock()
-	counts := make(map[string]int)
-	for _, c := range tg.gaugeCalls {
-		counts[c.name]++
-	}
-	for _, c := range tg.incrCalls {
-		counts[c.name]++
-	}
-	for _, c := range tg.countCalls {
-		counts[c.name]++
-	}
-	for _, c := range tg.timingCalls {
-		counts[c.name]++
-	}
-	return counts
-}
-
-func (tg *testStatsdClient) Counts() map[string]int64 {
-	tg.mu.RLock()
-	defer tg.mu.RUnlock()
-	c := make(map[string]int64)
-	for key, value := range tg.counts {
-		c[key] = value
-	}
-	return c
-}
-
-func (tg *testStatsdClient) Tags() []string {
-	tg.mu.RLock()
-	defer tg.mu.RUnlock()
-	t := make([]string, len(tg.tags))
-	copy(t, tg.tags)
-	return t
-}
-
-func (tg *testStatsdClient) Reset() {
-	tg.mu.Lock()
-	defer tg.mu.Unlock()
-	tg.gaugeCalls = tg.gaugeCalls[:0]
-	tg.incrCalls = tg.incrCalls[:0]
-	tg.countCalls = tg.countCalls[:0]
-	tg.timingCalls = tg.timingCalls[:0]
-	tg.counts = make(map[string]int64)
-	tg.tags = tg.tags[:0]
-	tg.n = 0
-}
-
-// Wait blocks until n metrics have been reported using the testStatsdClient or until duration d passes.
-// If d passes, or a wait is already active, an error is returned.
-func (tg *testStatsdClient) Wait(asserts *assert.Assertions, n int, d time.Duration) error {
-	c := func() bool {
-		tg.mu.RLock()
-		defer tg.mu.RUnlock()
-
-		return tg.n >= n
-	}
-	if !asserts.Eventually(c, d, 50*time.Millisecond) {
-		return fmt.Errorf("timed out after waiting %s for gauge events", d)
-	}
-
-	return nil
-}
-
 func TestReportRuntimeMetrics(t *testing.T) {
-	var tg testStatsdClient
+	var tg statsdtest.TestStatsdClient
 	trc := newUnstartedTracer(withStatsdClient(&tg))
 	defer trc.statsd.Close()
 
@@ -261,7 +44,7 @@ func TestReportRuntimeMetrics(t *testing.T) {
 
 func TestReportHealthMetrics(t *testing.T) {
 	assert := assert.New(t)
-	var tg testStatsdClient
+	var tg statsdtest.TestStatsdClient
 
 	defer func(old time.Duration) { statsInterval = old }(statsInterval)
 	statsInterval = time.Nanosecond
@@ -281,7 +64,7 @@ func TestReportHealthMetrics(t *testing.T) {
 
 func TestTracerMetrics(t *testing.T) {
 	assert := assert.New(t)
-	var tg testStatsdClient
+	var tg statsdtest.TestStatsdClient
 	tracer, _, flush, stop := startTestTracer(t, withStatsdClient(&tg))
 
 	tracer.StartSpan("operation").Finish()
@@ -296,11 +79,11 @@ func TestTracerMetrics(t *testing.T) {
 	assert.Equal(1, calls["datadog.tracer.flush_bytes"])
 	assert.Equal(1, calls["datadog.tracer.flush_traces"])
 	assert.Equal(int64(1), counts["datadog.tracer.flush_traces"])
-	assert.False(tg.closed)
+	assert.False(tg.Closed())
 
 	tracer.StartSpan("operation").Finish()
 	stop()
 	calls = tg.CallsByName()
 	assert.Equal(1, calls["datadog.tracer.stopped"])
-	assert.True(tg.closed)
+	assert.True(tg.Closed())
 }
